kernel: move handshake code to being manually triggered via devarch
The handshake code sends a RAS signal to the service node to tell it that
we are ready to receive ciod connections.  We were sending this in main
well before any user code was running.  This patch moves the handshake code
so it can be manually triggered during the initialization script after we
have setup a daemon to listen for ciod traffic.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>