brasil: add an extra export for csrv backmount
I got confused.  One export needs to be there to backmount csrv on Linux (we
use 9srv on Plan 9), the other export needs to be there to mount the host
file server when we aren't using csrv (which hopefully will only be temporary).
The user can set eithe rof these with command line arguments, but default is
5670 for host file system and 5640 for csrv backmount.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>