csrv-lite: non-automount verison of csrv
New verison of csrv-lite which doesn't attempt to work via
automount.  Instead, this version of csrv uses a ctl file which can
be interacted with a simple command language to mount and/or bind
resources into csrv mount points.

This initial version just implements two of the operational commands -
bind and mount which do simple (non-duct) manipulations of the namespace.
These combined with an existing export of the root file system should be
sufficient to give us base csrv functionality for taskfs.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>