brasil: different debug options
Added a -9 option to csrvlite which turns on 9P debugging under Plan 9
for the srv channel (allowing us to see all 9P traffic going via /csrv).
It might be nice in the future to put the styxmon someplace where the
other exports can see it as well so we get remote stuff.

Also made logging to a file no longer default so we get our 9P trace
interlaced with our error output.  To put it back to a log file, use -L.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>