anl-scripts: new simplified uplink
Added a simplified uplink script which can be used with Linux or MacOSX based
Inferno to establish an ssh tunnel to an ANL Inferno instance and export
that namespace over a conventional TCP/IP port.  Then 9vx, Plan 9 ports,
or a Plan 9 virtual machine can mount the ANL namespace and interact with it.
Included will be the local file system on surveyor, a /net capable of
accessing the BG/P, and a cmd(3) file system which can be used to launch
commands on the front-end nodes to allocate bumps or kick off runs.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>