mpipefs: userspace multipipe
I was having trouble debugging a kernel-space multipipe implementation,
so I re-implemented in userspace.  This version is a focused implementation
that will likely only work for very targeted applications.  At this point
it works like a normal pipe (with restrictions that write size can never
be greater than read size at the moment).

I'll be actually adding the multi portion of the multipipe tomorrow which
will also solve the aforementioned limitation.

Signed-off-by: Eric Van Hensbergen <ericvh@gmail.com>