added fillstatbuf to quiet down the gangfs sync communication
added LOTS of debugging

fixed problems with the parent path definition

Fixed a problem where checkmount checked for the existence of the
mount point where it was already created by another process (and could
therefor see it), but it was not mounted in the local namespace.

added releasegang and cleanupgang to fswrite after completion of the
cmdbcast.

fixed a string termination error which caused the buffer returned by
pread to be unterminated and caused the proc ID return to be invalid.

added amprefix in place of "/n" for consistency.

fixed the execfs session path to be execfspath instead of procpath.